+++ apps/admin-api/src/common/errors/BadRequestError.ts
@@ -1,0 +1,11 @@
+import { GraphQLError } from 'graphql'
+
+export class BadRequestError extends GraphQLError {
+  constructor(message: string) {
+    super(message, {
+      extensions: {
+        code: 'BAD_REQUEST',
+      },
+    })
+  }
+}