+++ apps/admin-api/src/graphql.ts
@@ -1,78 +1,20 @@
-/** ------------------------------------------------------
+/*
+ * -------------------------------------------------------
  * THIS FILE WAS AUTOMATICALLY GENERATED (DO NOT MODIFY)
  * -------------------------------------------------------
  */
 
 /* tslint:disable */
 /* eslint-disable */
-export enum DayOfWeek {
-  MO = 'MO',
-  TU = 'TU',
-  WE = 'WE',
-  TH = 'TH',
-  FR = 'FR',
-  SA = 'SA',
-  SU = 'SU',
-  IA = 'IA',
-  AR = 'AR',
-}
-
-export enum StudyProgram {
-  S = 'S',
-  T = 'T',
-  I = 'I',
-}
-
-export enum GenEdType {
-  SO = 'SO',
-  HU = 'HU',
-  SC = 'SC',
-  IN = 'IN',
-  NO = 'NO',
-}
-
-export enum ReviewInteractionType {
-  L = 'L',
-  D = 'D',
-}
-
-export enum ReviewStatus {
-  APPROVED = 'APPROVED',
-  REJECTED = 'REJECTED',
-  PENDING = 'PENDING',
-}
-
-export class PeriodRangeInput {
-  start: string
-  end: string
-}
-
-export class FilterInput {
-  keyword?: string
-  genEdTypes?: GenEdType[]
-  dayOfWeeks?: DayOfWeek[]
-  periodRange?: PeriodRangeInput
-  limit?: number
-  offset?: number
-}
-
-export class CourseGroupInput {
-  semester: string
-  academicYear: string
-  studyProgram: StudyProgram
-}
-
-export class GenEdOverrideInput {
-  genEdType: GenEdType
-  sections: string[]
-}
+export type DayOfWeek = 'MO' | 'TU' | 'WE' | 'TH' | 'FR' | 'SA' | 'SU' | 'IA' | 'AR'
+export type StudyProgram = 'S' | 'T' | 'I'
+export type GenEdType = 'SO' | 'HU' | 'SC' | 'IN' | 'NO'
+export type ReviewInteractionType = 'L' | 'D'
+export type ReviewStatus = 'APPROVED' | 'REJECTED' | 'PENDING'
 
 export class OverrideInput {
   courseNo: string
-  studyProgram: StudyProgram
-  semester: string
-  academicYear: string
-  genEd?: GenEdOverrideInput
+  genEdType: GenEdType
 }
 
 export class CreateReviewInput {
@@ -81,85 +23,22 @@ export class CreateReviewInput {
   semester: string
   academicYear: string
   studyProgram: StudyProgram
-  content?: string
+  content?: Nullable<string>
 }
 
 export class EditReviewInput {
-  rating?: number
-  semester?: string
-  academicYear?: string
-  content?: string
+  rating?: Nullable<number>
+  semester?: Nullable<string>
+  academicYear?: Nullable<string>
+  content?: Nullable<string>
 }
 
-export class Period {
-  start: string
-  end: string
-}
-
-export class ExamPeriod {
-  date: string
-  period: Period
-}
-
-export class Capacity {
-  current: number
-  max: number
-}
-
-export class Class {
-  type: string
-  dayOfWeek?: DayOfWeek
-  period?: Period
-  building?: string
-  room?: string
-  teachers: string[]
-}
-
-export class Section {
-  sectionNo: string
-  closed: boolean
-  capacity: Capacity
-  note?: string
-  classes: Class[]
-  genEdType: GenEdType
-}
-
-export class Course {
-  studyProgram: StudyProgram
-  semester: string
-  academicYear: string
+export class Override {
   courseNo: string
-  courseDescTh?: string
-  courseDescEn?: string
-  abbrName: string
-  courseNameTh: string
-  courseNameEn: string
-  faculty: string
-  department: string
-  credit: number
-  creditHours: string
-  courseCondition: string
   genEdType: GenEdType
-  midterm?: ExamPeriod
-  final?: ExamPeriod
-  sections: Section[]
-  rating?: string
-  courseDesc?: string
-}
-
-export class CourseNosOutput {
-  S: string[]
-  T: string[]
-  I: string[]
 }
 
 export abstract class IQuery {
-  abstract courseNos(): CourseNosOutput | Promise<CourseNosOutput>
-
-  abstract course(courseNo: string, courseGroup: CourseGroupInput): Course | Promise<Course>
-
-  abstract search(filter: FilterInput, courseGroup: CourseGroupInput): Course[] | Promise<Course[]>
-
   abstract overrides(): Override[] | Promise<Override[]>
 
   abstract reviews(courseNo: string, studyProgram: StudyProgram): Review[] | Promise<Review[]>
@@ -172,26 +51,10 @@ export abstract class IQuery {
   abstract pendingReviews(): Review[] | Promise<Review[]>
 }
 
-export class GenEdOverride {
-  genEdType: GenEdType
-  sections: string[]
-}
-
-export class Override {
-  courseNo: string
-  studyProgram: StudyProgram
-  semester: string
-  academicYear: string
-  genEd?: GenEdOverride
-}
-
 export abstract class IMutation {
   abstract createOrUpdateOverride(override: OverrideInput): Override | Promise<Override>
 
-  abstract deleteOverride(
-    courseNo: string,
-    courseGroup: CourseGroupInput
-  ): Override | Promise<Override>
+  abstract deleteOverride(courseNo: string): Nullable<Override> | Promise<Nullable<Override>>
 
   abstract createReview(createReviewInput: CreateReviewInput): Review | Promise<Review>
 
@@ -207,7 +70,7 @@ export abstract class IMutation {
   abstract setReviewStatus(
     reviewId: string,
     status: ReviewStatus,
-    rejectionReason?: string
+    rejectionReason?: Nullable<string>
   ): string | Promise<string>
 }
 
@@ -215,14 +78,17 @@ export class Review {
   _id: string
   rating: number
   courseNo: string
+  courseTitle: string
   semester: string
   academicYear: string
   studyProgram: StudyProgram
-  content?: string
+  content?: Nullable<string>
   likeCount: number
   dislikeCount: number
-  myInteraction?: ReviewInteractionType
-  status?: ReviewStatus
-  rejectionReason?: string
+  myInteraction?: Nullable<ReviewInteractionType>
+  status?: Nullable<ReviewStatus>
+  rejectionReason?: Nullable<string>
   isOwner: boolean
 }
+
+type Nullable<T> = T | null