+++ apps/admin-web/src/common/Topbar/styled.ts
@@ -1,6 +1,0 @@
-import { AppBar, styled } from '@mui/material'
-
-export const StyledAppBar = styled(AppBar)({
-  backgroundColor: 'black',
-  padding: '15px 10px',
-})