+++ apps/admin-web/src/context/AppProvider.tsx
@@ -1,9 +1,12 @@
 import { ApolloProvider } from '@apollo/client'
 import { ThemeProvider } from '@mui/material'
 
-import Topbar from '@admin-web/common/Topbar/Topbar'
-import { defaultTheme } from '@admin-web/config/theme/index'
+import { Layout } from '@admin-web/common/Layout'
+import { defaultTheme } from '@admin-web/config/theme'
 import { client } from '@admin-web/services/apollo'
+
+import AuthProvider from './AuthProvider'
+import { ProtectedRoutes } from './ProtectedRoutes'
 
 interface AppProviderProps {
   children: React.ReactNode
@@ -12,10 +15,13 @@ interface AppProviderProps {
 export function AppProvider({ children }: AppProviderProps) {
   return (
     <ApolloProvider client={client}>
-      <ThemeProvider theme={defaultTheme}>
-        <Topbar />
-        {children}
-      </ThemeProvider>
+      <AuthProvider>
+        <ThemeProvider theme={defaultTheme}>
+          <Layout>
+            <ProtectedRoutes>{children}</ProtectedRoutes>
+          </Layout>
+        </ThemeProvider>
+      </AuthProvider>
     </ApolloProvider>
   )
 }