+++ apps/admin-web/src/module/genEd/components/AddGenEdDialog/styled.tsx
@@ -1,0 +1,16 @@
+import { styled } from '@mui/material'
+
+export const DialogContainer = styled('div')({
+  display: 'flex',
+  flexDirection: 'column',
+  alignContent: 'start',
+  rowGap: '16px',
+  padding: '24px',
+  background: '#ECFEFF',
+})
+
+export const ButtonContainer = styled('div')({
+  display: 'flex',
+  alignContent: 'center',
+  columnGap: '16px',
+})