+++ apps/admin-web/src/pages/generateToken/index.tsx
@@ -1,0 +1,1 @@
+export { GenerateToken as default } from '@admin-web/module/generateToken'