+++ apps/admin-web/src/services/apollo/index.ts
@@ -2,16 +2,19 @@ import { ApolloClient, InMemoryCache } from '@apollo/client'
 import { BatchHttpLink } from '@apollo/client/link/batch-http'
 import { setContext } from '@apollo/client/link/context'
 
-import { ADMIN_ACCESS_TOKEN, ENVIRONMENT } from '@admin-web/env'
+import { ENVIRONMENT } from '@admin-web/env'
 import { apiUrl } from '@admin-web/services/httpClient'
 
 const createHttpLink = () =>
   new BatchHttpLink({
     uri: `${apiUrl}/graphql`,
+
+    // For cross-site cookie
+    credentials: 'include',
   })
 
 const authLink = setContext(async (_, { headers }) => {
-  const accessToken = ADMIN_ACCESS_TOKEN
+  const accessToken = 'ADMIN_ACCESS_TOKEN'
   return {
     headers: {
       ...headers,
@@ -24,6 +27,11 @@ export const client = new ApolloClient({
   link: authLink.concat(createHttpLink()),
   cache: new InMemoryCache(),
   connectToDevTools: ENVIRONMENT !== 'production',
+  defaultOptions: {
+    mutate: {
+      errorPolicy: 'all',
+    },
+  },
 })
 
 export function createApolloServerClient() {