+++ apps/admin-api/CHANGELOG.md
@@ -1,5 +1,21 @@
 # admin-api
 
+## 1.0.0-beta.1
+
+### Major Changes
+
+- f771ad38: init admin-api and admin-web
+
+## 0.2.0-beta.0
+
+### Minor Changes
+
+- 79af67e9: bump mongoose to 6.12 (fix vulnerabilities)
+
+### Patch Changes
+
+- 8d9d304b: bump eslint and fix ci
+
 ## 0.1.1
 
 ### Patch Changes