+++ apps/web/CHANGELOG.md
@@ -1,5 +1,20 @@
 # web
 
+## 1.7.0-beta.1
+
+### Minor Changes
+
+- 46e520b2: feat: explicit warning not reg chula at first page (Close #623)
+
+## 1.6.1-beta.0
+
+### Patch Changes
+
+- 8d9d304b: bump eslint and fix ci
+- Updated dependencies [8d9d304b]
+  - @cgr/course-utils@1.0.3-beta.0
+  - @cgr/codegen@1.2.1-beta.0
+
 ## 1.6.0
 
 ### Minor Changes